% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lncRNA_degree.R
\name{compute_lncRNA_degree}
\alias{compute_lncRNA_degree}
\title{Compute Connectivity (Degree) of lncRNAs in Each Layer}
\usage{
compute_lncRNA_degree(
  net,
  genes_list,
  layers,
  mode = "all",
  output_file = "lncRNAs-degrees_bylayer.csv"
)
}
\arguments{
\item{net}{A multilayer network object of class \code{ml.network}, typically created using
\code{create_multilayer_network()}.}

\item{genes_list}{A character vector of lncRNA gene names to analyze.}

\item{layers}{A character vector of layer names.}

\item{mode}{Degree mode for multilayer network (default = "all").}

\item{output_file}{File name to save the wide-format result table. Default = \code{"lncRNAs-degrees_bylayer.csv"}.}
}
\value{
A data frame in wide format with layers as rows and lncRNAs as columns.
}
\description{
This function calculates the degree (number of connections) for each lncRNA
in every layer of a multilayer network.
}
\examples{
\dontrun{
lncRNAs <- c("AC011483.1", "DSG1-AS1", "TMEM99")
layers <- layers_ml(net)
compute_lncRNA_degree(net, genes_list = lncRNAs, layers = layers)
}

}
