% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene_types_from_net.R
\name{extract_gene_types_from_net}
\alias{extract_gene_types_from_net}
\title{Extract GeneType Attributes from a Multilayer Network}
\usage{
extract_gene_types_from_net(net)
}
\arguments{
\item{net}{A multilayer network object of class \code{ml.network}, typically
created using \code{create_multilayer_network()} and enriched using
\code{add_network_attributes()}.}
}
\value{
A data frame with two columns:
\describe{
\item{GeneName}{The name of each actor (gene) in the network.}
\item{GeneType}{The gene type assigned to each actor.}
}
}
\description{
This function retrieves the \code{GeneType} attribute for all actors (genes)
present in a multilayer network object created with the \code{multinet} package.
}
\details{
This function uses \code{actors_ml()} and \code{get_values_ml()} to extract actor names
and their assigned GeneType attributes.
}
\examples{
\dontrun{
actors_with_type <- extract_gene_types_from_net(net)
}

}
