% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adjacency_matrices.R
\name{generate_adjacency_matrices}
\alias{generate_adjacency_matrices}
\title{Generate Adjacency Matrices Based on Gene Expression and Age Groups}
\usage{
generate_adjacency_matrices(
  threshold = 0.7,
  num_repetitions = 50,
  num_samples = 10,
  correlation = "spearman",
  expression_mat,
  metaData,
  genesInfo,
  TargetGenes
)
}
\arguments{
\item{threshold}{Numeric. Correlation threshold below which values are set to 0. Default is 0.7.}

\item{num_repetitions}{Integer. Number of repetitions for sampling and computing adjacency matrices. Default is 50.}

\item{num_samples}{Integer. Number of samples to randomly select per group for each repetition. Default is 10.}

\item{correlation}{Character. Type of correlation to compute: \code{"spearman"} or \code{"pearson"}. Default is \code{"spearman"}.}

\item{expression_mat}{A matrix or data frame of normalized gene expression data. Genes are in rows, samples in columns.}

\item{metaData}{A data frame containing metadata with at least two columns: \code{"Groups"} and \code{"Samples"} (sample names).}

\item{genesInfo}{A data frame containing information about each gene, with at least two columns; \code{"GeneName"} and \code{"GeneType"}.}

\item{TargetGenes}{A data frame of differentially expressed genes with at least a \code{"DEgenes"} column.}
}
\value{
A list of length \code{num_repetitions}, where each element is a list of adjacency matrices (one per age group).
Each adjacency matrix is a weighted, symmetric matrix of gene-gene correlations after thresholding.
}
\description{
This function creates weighted adjacency matrices for each group
across multiple repetitions by computing pairwise gene-gene correlations
from gene expression data. It filters out low-correlation values based on
a specified threshold. Moreover, it is using goodSamplesGenes method to select good actors.
}
\examples{
\dontrun{
adjacency_matrices <- generate_adjacency_matrices(
  expression_mat = my_expression_matrix,
  metaData = my_meta,
  genesInfo = my_genes_info,
  TargetGenes = my_deg_list,
  correlation = "spearman",
  threshold = 0.7,
  num_repetitions = 10,
  num_samples = 5
)
}

}
