% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_network_attributes.R
\name{add_network_attributes}
\alias{add_network_attributes}
\title{Add Node-Level Attributes to a Multilayer Network}
\usage{
add_network_attributes(net, genes_info)
}
\arguments{
\item{net}{A multilayer network object of class \code{ml.network}, created using
functions such as create_multilayer_network().}

\item{genes_info}{A data frame containing gene metadata. It must include at least
the columns \code{GeneName} and \code{GeneType}.}
}
\value{
This function updates the \code{net} object in-place by assigning the \code{GeneType}
attribute to all actors. It does not return a value.
}
\description{
This function adds gene-level metadata (e.g., GeneType) to actors (nodes)
in a multilayer network constructed using the \code{multinet} package.
}
\details{
\itemize{
\item Extracts actors from the network
\item Joins metadata
\item Fills missing values with "Unknown"
\item Assigns \code{GeneType} as an actor attribute
}
}
\examples{
\dontrun{
add_network_attributes(net, genes_info)
}

}
